requirements.txt
PyMuPDF==1.23.26
Pillow==10.1.0
requests==2.31.0
openai==1.3.7
lxml==4.9.3
pdf2image==1.16.3
reportlab==4.0.7
python-dotenv==1.0.0